Use a square collider on ball to avoid loss of momentum
The sphere collider does not seem to play well with the angles of the
blocks when using Unity's build in physics.

Instead of rolling my own solution, I would rather have the object
behaving as I expect it to, and thus I'm going with the square
collider.. for now.

This may change in the future.